+++ CHANGELOG.md
@@ -10,7 +10,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - chore: raise minimal Python version to 3.12.
 - fix: improve plotting of 2d matrices (detector effects).
 - feature: add dynamic versioning.
-- feature: add switcher `is_global_normalization_free` to control state of `global_normalization` parameter. Default state of it is fixed.
+- feature: add switcher `is_global_normalization_fixed` to control state of `global_normalization` parameter. Default state of it is fixed.
 
 ## [0.7.1] - 2026-01-29